+++ custom_components/petlibro/devices/feeders/one_rfid_smart_feeder.py
@@ -229,7 +229,7 @@ class OneRFIDSmartFeeder(Device):
         return None
     
     @property
-    def last_feed_quantity(self) -> float | None:
+    def last_feed_quantity(self) -> int | None:
         """Return the last feed amount in raw grain count."""
         raw = self._data.get("workRecord", [])
         if not raw or not isinstance(raw, list):
